+++ magik-squid/src/main/java/org/stevenlooman/sw/magik/api/MagikKeyword.java
@@ -43,6 +43,7 @@ public enum MagikKeyword implements GrammarRuleKey {
   SUPER,
   PRIMITIVE,
   THISTHREAD,
+  CLASS,
   ;
 
   /**