+++ CHANGES.md
@@ -4,6 +4,7 @@ Changes
 0.1.1 (unreleased)
 
 - Add --untabify <n> option to magik-lint
+- Add support for \_class keyword
 
 
 0.1.0 (2019-06-19)